pyproject.toml
[tool.poetry]
name = "individual-creative-task-os"
version = "0.1.0"
description = ""
authors = ["DeLight <nazarchukyayko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
pygame = "^2.6.1"
pyassimp = "^5.2.5"
pyopengl = "^3.1.7"
setuptools = "^75.5.0"
assimp = "^4.1.4"
pyfbx = "^0.0.8"
numpy = "^2.1.3"
glm = "^0.4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
